pyproject.toml
[project]
name = "s10-mini-capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.0",
    "google-genai>=1.52.0",
    "markitdown>=0.1.3",
    "mcp[cli]>=1.22.0",
    "pillow>=12.0.0",
    "pymupdf4llm>=0.2.4",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
]
